[CHORE] Double 자료형 수정

+++ 30th-Assignment/30th-Assignment/30th-Assignment/Global/Extension/Adjusted+.swift
@@ -27,12 +27,12 @@ extension CGFloat {
 
 extension Double {
     var adjustedWidth: Double {
-        let ratio: Double = UIScreen.main.bounds.width / 375
+        let ratio: Double = Double(UIScreen.main.bounds.width / 375)
         return self * ratio
     }
 
     var adjustedHeight: Double {
-        let ratio: Double = UIScreen.main.bounds.height / 812
+        let ratio: Double = Double(UIScreen.main.bounds.height / 812)
         return self * ratio
     }
 }
